The SunOS' egrep fails to process properly the `egrep' invocations of the test suite. * tests/aclocal.m4 (AC_STATE_SAVE): If egrep fails, remove the output file. * tests/atspecific.m4 (_AT_CHECK_AC_MACRO): Don't check `state-*' if the files are not present. * m4sugar.m4 (m4_join): Rename as... (m4_smash): this. * tests/aclocal.m4 (join): Move as... * m4sugar.m4 (m4_flatten): this. * autoconf.sh (trace.m4): Rename m4_smash as m4_flatten.

+++ tests/atspecific.m4
@@ -72,9 +72,11 @@ AT_CHECK_CONFIGURE
 
 dnl Some tests might exit prematurely when they find a problem, in
 dnl which case `env-after' is probably missing.  Don't check it then.
-if test -f state-env.after; then
+if test -f state-env.before -a -f state-env.after; then
   mv -f state-env.before expout
   AT_CHECK([cat state-env.after], 0, expout)
+fi
+if test -f state-ls.before -a -f state-ls.after; then
   mv -f state-ls.before expout
   AT_CHECK([cat state-ls.after], 0, expout)
 fi
